refactor: 불필요한 useChannel 훅 제거
